add default completion if there is no result or no exact match, render it in dim color
in this way, we don't need to worry about fuzzy result is aggressive that user cannot create a person
with a similar name to an existing one. even if a very similar one is matched, user can still move down
to select the suggestion with dimmed color and "(new)" suffix to create the person.

thus threshold can be increased a bit.